Fix Email Theme (#3)
The email theme isn't correctly configured to use the Keycloak theme as
a base. This causes it to break for any template we don't have defined
here. This commit overhauls the email theme to fix the issues, and
greatly reduce the amount of code to maintain.

It also removes the messages for the login theme, since we don't seem to
have any overrides for the login messages.

* Set keycloak theme as base for email theme
* Remove email templates - they're just copies of the base templates
* Reduce email messages to just the overrides
* Remove the login message overrides - there weren't any overrides.